Extract sign-in capabilities into dedicated GET /flow/sign-in API
Move the sign-in capability description (password_sign_in, identifier
claims, providers) out of GET /flow/configuration into a new dedicated
GET /flow/sign-in endpoint, so the "what sign-in methods exist" contract
is owned by the sign-in API instead of the generic flow-configuration
blob and can be evolved/reused independently.

- Add SignInConfigurationResource and relocate PasswordConfigurationResource
  and ProviderConfigurationResource next to it.
- Add GET handler + getPassword()/getProvider() to SignInController.
- Remove password/providers from ConfigurationResource and password_sign_in
  from FeaturesResource; simplify ConfigurationController accordingly.
- password is now null when password sign-in is disabled, matching its
  schema doc.

BREAKING: the front-end must read password/providers/password_sign_in from
GET /flow/sign-in instead of GET /flow/configuration.
