+++ server/src/main/kotlin/com/sympauthy/api/controller/flow/ConfigurationController.kt
@@ -1,18 +1,10 @@
 package com.sympauthy.api.controller.flow
 
-import com.sympauthy.api.controller.flow.ProvidersController.Companion.FLOW_PROVIDER_AUTHORIZE_ENDPOINT
-import com.sympauthy.api.controller.flow.ProvidersController.Companion.FLOW_PROVIDER_ENDPOINTS
 import com.sympauthy.api.controller.flow.util.WebAuthorizationFlowControllerUtil
 import com.sympauthy.api.resource.flow.*
-import com.sympauthy.business.manager.ClientManager
 import com.sympauthy.business.manager.ClaimManager
+import com.sympauthy.business.manager.ClientManager
 import com.sympauthy.business.manager.flow.WebAuthorizationFlowPasswordManager
-import com.sympauthy.business.manager.provider.ProviderManager
-import com.sympauthy.business.model.provider.EnabledProvider
-import com.sympauthy.config.model.EnabledUrlsConfig
-import com.sympauthy.config.model.UrlsConfig
-import com.sympauthy.config.model.getUri
-import com.sympauthy.config.model.orThrow
 import com.sympauthy.security.SecurityRule.HAS_STATE
 import com.sympauthy.security.stateOrNull
 import com.sympauthy.server.DisplayMessages
@@ -25,8 +17,6 @@ import io.micronaut.security.annotation.Secured
 import io.micronaut.security.authentication.Authentication
 import io.swagger.v3.oas.annotations.Operation
 import jakarta.inject.Inject
-import kotlinx.coroutines.async
-import kotlinx.coroutines.coroutineScope
 import java.util.*
 
 @Secured(HAS_STATE)
@@ -35,9 +25,7 @@ class ConfigurationController(
     @Inject private val claimManager: ClaimManager,
     @Inject private val clientManager: ClientManager,
     @Inject private val passwordFlowManager: WebAuthorizationFlowPasswordManager,
-    @Inject private val providerManager: ProviderManager,
     @Inject private val webAuthorizationFlowControllerUtil: WebAuthorizationFlowControllerUtil,
-    @Inject private val uncheckedUrlsConfig: UrlsConfig,
     @Inject @param:DisplayMessages private val displayMessageSource: MessageSource
 ) {
 
@@ -45,7 +33,7 @@ class ConfigurationController(
         description = """
 Return the configuration of the authentication flow associated to the client. It exposes information such as:
 - features that are enabled for the client.
-- third party providers that can be used to authenticate the end-user.
+- claims collectable by the authorization server.
 - etc.
 
 This configuration only contains information that are associated to this authorization server and the client,
@@ -60,30 +48,15 @@ they can be cached across the different end-users trying to authenticate.
     ): ConfigurationResource = webAuthorizationFlowControllerUtil.fetchOnGoingAttemptThenRun(
         state = authentication.stateOrNull
     ) { authorizeAttempt, _ ->
-        coroutineScope {
-            val locale = httpRequest.locale.orDefault()
-            val urlsConfig = uncheckedUrlsConfig.orThrow()
+        val locale = httpRequest.locale.orDefault()
 
-            val client = clientManager.findClientByIdOrNull(authorizeAttempt.clientId)
-            val audience = client?.audience
+        val client = clientManager.findClientByIdOrNull(authorizeAttempt.clientId)
+        val audience = client?.audience
 
-            val deferredClaims = async {
-                getCollectableClaims(locale)
-            }
-            val features = getFeatures(audience)
-            val deferredProviders = async {
-                providerManager.listEnabledProviders()
-                    .takeIf(List<EnabledProvider>::isNotEmpty)
-                    ?.map { getProvider(it, urlsConfig) }
-            }
-
-            ConfigurationResource(
-                claims = deferredClaims.await(),
-                features = features,
-                password = getPassword(),
-                providers = deferredProviders.await()
-            )
-        }
+        ConfigurationResource(
+            claims = getCollectableClaims(locale),
+            features = getFeatures(audience)
+        )
     }
 
     private fun getCollectableClaims(locale: Locale): List<CollectableClaimConfigurationResource> {
@@ -102,31 +75,9 @@ they can be cached across the different end-users trying to authenticate.
         audience: com.sympauthy.business.model.audience.Audience?
     ): FeaturesResource {
         return FeaturesResource(
-            passwordSignIn = passwordFlowManager.signInEnabled,
             signUp = passwordFlowManager.signUpEnabled,
             signUpEnabled = audience?.signUpEnabled ?: true,
             invitationEnabled = audience?.invitationEnabled ?: false,
-        )
-    }
-
-    private fun getPassword(): PasswordConfigurationResource {
-        return PasswordConfigurationResource(
-            identifierClaims = claimManager.listIdentifierClaims().map { it.id }
-        )
-    }
-
-    private fun getProvider(
-        provider: EnabledProvider,
-        urlsConfig: EnabledUrlsConfig
-    ): ProviderConfigurationResource {
-        val authorizeUrl = urlsConfig.getUri(
-            FLOW_PROVIDER_ENDPOINTS + FLOW_PROVIDER_AUTHORIZE_ENDPOINT,
-            "providerId" to provider.id
-        )
-        return ProviderConfigurationResource(
-            id = provider.id,
-            name = provider.name,
-            authorizeUrl = authorizeUrl.toString()
         )
     }
 }
